Refactor telemetry system and improve UART communication
- Added telemetry functionality with structured packet format for sending flight data.
- Implemented DMA for USART1 to enhance data transmission efficiency.
- Adjusted UART baud rate to 460800 for higher data throughput.
- Introduced TIM5 for accurate timing and delta time calculations.
- Updated motor control functions to use float instead of double for consistency.
- Modified SPI settings for improved communication with nRF24L01+ module.
- Enhanced error handling and added callbacks for UART transmission completion.
- Cleaned up commented-out code and improved overall code readability.

+++ flight_C/Core/Inc/IMU.h
@@ -4,12 +4,12 @@
 #include "main.h"
 
 typedef struct {
-    double pitch;
-    double roll;
-    double yaw;
-    double yaw_rate;
-    double pitch_rate;
-    double roll_rate;
+    float pitch;
+    float roll;
+    float yaw;
+    float yaw_rate;
+    float pitch_rate;
+    float roll_rate;
 } IMUData;
 
 void    IMU_Init(void);
@@ -17,7 +17,7 @@ void    MPU6050_Init(void);
 void    MPU6050_Read(int16_t *ax, int16_t *ay, int16_t *az,
                      int16_t *gx, int16_t *gy, int16_t *gz);
 void    IMU_Calibrate(int samples);
-IMUData IMU_GetAngles(double dt);
+IMUData IMU_GetAngles(float dt);
 void    IMU_PrintTelemetry(IMUData data);
